Enhance code and examples (version 1.1)
The examples were previously incorrect as they seem to have been a
copy-paste of heart canisters and still had heart canister ID despite
checking for example IDs. That has been fixed.

Additionally the changes that are being introduced in GM4_Datapacks
are included in this commit. Other than some refactoring, the primary
change is that `SkullOwner.Name` is now compared against the `name` tag
within the heads registry as an alternative to comparing the texture
tag against `value`. This allows different skulls to be differentiated
even if they share the same texture.

Additionally, as a minor tweak, an item with the tag
`gm4_player_head:1b` will be excluded from the selector in the `main`
function, and therefore the item will not be repaired. The presence of
this NBT tag implies that the skull was created by commands rather than
a placed/broken skull, so repair is unnecessary (or potentially harmful,
in the case that a broken item is converted into an "inert" form upon
being placed or broken).